v0.6.44: 주간보고 DB·페이지, 진척률 numeric·계획 진행,보내기 기본 프로젝트
Supabase 주간보고·작성 UI, tasks.progress 소수 저장 및 계획 진행 오버라이드,
엑셀·WBS 표·선택 훅 보완, 전체 보기에서도보내기 시 프로젝트가 기본 선택되도록 함.

+++ supabase/migrations/FULL_SETUP_NEW_PROJECT.sql
@@ -21,7 +21,7 @@ CREATE TABLE IF NOT EXISTS tasks (
   name text NOT NULL,
   start_date date NOT NULL,
   end_date date NOT NULL,
-  progress integer DEFAULT 0,
+  progress numeric(6, 2) DEFAULT 0,
   assignee text DEFAULT '',
   status text DEFAULT 'todo',
   expanded boolean DEFAULT false,
